% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app_code.R
\name{mapDirectory}
\alias{mapDirectory}
\title{Map directory structure}
\usage{
mapDirectory(dir = ".", drop_root = FALSE, include_files = TRUE)
}
\arguments{
\item{dir}{\code{\link[base]{character}}.
Directory path.}

\item{drop_root}{\code{\link[base]{logical}}.
\code{TRUE}: include root directory;
\code{FALSE}: do not include root directory. See examples.}

\item{include_files}{\code{\link[base]{logical}}.
\code{TRUE}: include actual files;
\code{FALSE}: do not include files. See examples.}
}
\value{
\code{\link[base]{list}}.
}
\description{
Maps directory structure to a \code{\link{list}}.
}
\details{
Whether or not only directories or actual files should be included can
be controlled by \code{include_files}.
}
\examples{
mapDirectory()
mapDirectory(drop_root = TRUE)
mapDirectory(include_files = TRUE)


}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/shinyref}
}

