% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinydev.r
\docType{package}
\name{shinydev}
\alias{shinydev}
\alias{shinydev-package}
\title{Package Title}
\description{
Package description
}
\details{
The core functions/methods of this package:
 \itemize{
   \item{\code{\link[shinydev]{function-name}}: }{

     Description of this function.

     Use empty lines for creating vertical space.
   }
   \item{\code{\link[shinydev]{function-name}}: }{

     Description of this function.

     Use empty lines for creating vertical space.
   }
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/shinyref}
}

